+++ app/js/modules/index.js
@@ -2,7 +2,7 @@
   'use strict';
 
   var SQX = global.SQX = global.SQX || {};
-  SQX.moduleOrder = SQX.moduleOrder || ['core', 'config', 'storage', 'license', 'ui', 'formatters', 'domain', 'datasets', 'renderers', 'charts', 'strategies', 'home', 'workflow', 'project-generator'];
+  SQX.moduleOrder = SQX.moduleOrder || ['core', 'config', 'storage', 'license', 'ui', 'formatters', 'domain', 'datasets', 'renderers', 'charts', 'strategies', 'home', 'support', 'fulfillment', 'customer-cockpit', 'workflow', 'project-generator'];
 
   SQX.boot = SQX.boot || function bootSQXModules() {
     SQX.booted = true;